Add non-generic DistributedApplicationFactory
